moved type converters into seperate class
